+++ lib/src/math/sp3d_constant_values.dart
@@ -1,0 +1,13 @@
+import 'dart:math';
+
+/// (en)This class is a collection of constants useful when using the simple_3d package.
+///
+/// (ja)simple_3dパッケージを利用する際に便利な定数をまとめたクラスです。
+///
+/// Author Masahide Mori
+///
+/// First edition creation date 2023-09-10 14:41:50
+class Sp3dConstantValues {
+  /// Constant to multiply when converting angles to radians.
+  static const double toRadian = pi / 180;
+}