+++ lib/src/util_sp3d_common_parts.dart
@@ -1,5 +1,4 @@
 import 'package:simple_3d/simple_3d.dart';
-import 'dart:math';
 import '../util_simple_3d.dart';
 
 /// (en)A utility to generate common parts using sp3d.
@@ -10,8 +9,6 @@ import '../util_simple_3d.dart';
 /// First edition creation date 2023-01-01 21:10:36
 ///
 class UtilSp3dCommonParts {
-  static const double _toRadian = pi / 180;
-
   /// (en)Generates arrows representing x,y,z coordinates.
   ///
   /// (ja)x、y、z 座標を表す矢印を生成します。
@@ -46,10 +43,10 @@ class UtilSp3dCommonParts {
     if (isCube) {
       Sp3dObj xAxisPillar = UtilSp3dGeometry.cube(r * 2, r * 2, h, 1, 1, 1,
           material: materialX ?? FSp3dMaterial.redNonWire.deepCopy());
-      xAxisPillar.rotate(Sp3dV3D(0, 1, 0), 90 * _toRadian);
+      xAxisPillar.rotate(Sp3dV3D(0, 1, 0), 90 * Sp3dConstantValues.toRadian);
       Sp3dObj yAxisPillar = UtilSp3dGeometry.cube(r * 2, r * 2, h, 1, 1, 1,
           material: materialY ?? FSp3dMaterial.greenNonWire.deepCopy());
-      yAxisPillar.rotate(Sp3dV3D(1, 0, 0), -90 * _toRadian);
+      yAxisPillar.rotate(Sp3dV3D(1, 0, 0), -90 * Sp3dConstantValues.toRadian);
       Sp3dObj zAxisPillar = UtilSp3dGeometry.cube(r * 2, r * 2, h, 1, 1, 1,
           material: materialZ ?? FSp3dMaterial.blueNonWire.deepCopy());
       objList.add(xAxisPillar);
@@ -58,10 +55,10 @@ class UtilSp3dCommonParts {
     } else {
       Sp3dObj xAxisPillar = UtilSp3dGeometry.pillar(r, r, h,
           material: materialX ?? FSp3dMaterial.redNonWire.deepCopy());
-      xAxisPillar.rotate(Sp3dV3D(0, 1, 0), 90 * _toRadian);
+      xAxisPillar.rotate(Sp3dV3D(0, 1, 0), 90 * Sp3dConstantValues.toRadian);
       Sp3dObj yAxisPillar = UtilSp3dGeometry.pillar(r, r, h,
           material: materialY ?? FSp3dMaterial.greenNonWire.deepCopy());
-      yAxisPillar.rotate(Sp3dV3D(1, 0, 0), -90 * _toRadian);
+      yAxisPillar.rotate(Sp3dV3D(1, 0, 0), -90 * Sp3dConstantValues.toRadian);
       Sp3dObj zAxisPillar = UtilSp3dGeometry.pillar(r, r, h,
           material: materialZ ?? FSp3dMaterial.blueNonWire.deepCopy());
       objList.add(xAxisPillar);
@@ -72,11 +69,11 @@ class UtilSp3dCommonParts {
     if (useArrowHead) {
       Sp3dObj xAxisArrow = UtilSp3dGeometry.cone(arrowHeadR, arrowHeadH,
           material: materialX ?? FSp3dMaterial.redNonWire.deepCopy());
-      xAxisArrow.rotate(Sp3dV3D(0, 1, 0), 90 * _toRadian);
+      xAxisArrow.rotate(Sp3dV3D(0, 1, 0), 90 * Sp3dConstantValues.toRadian);
       xAxisArrow.move(Sp3dV3D(h, 0, 0));
       Sp3dObj yAxisArrow = UtilSp3dGeometry.cone(arrowHeadR, arrowHeadH,
           material: materialY ?? FSp3dMaterial.greenNonWire.deepCopy());
-      yAxisArrow.rotate(Sp3dV3D(1, 0, 0), -90 * _toRadian);
+      yAxisArrow.rotate(Sp3dV3D(1, 0, 0), -90 * Sp3dConstantValues.toRadian);
       yAxisArrow.move(Sp3dV3D(0, h, 0));
       Sp3dObj zAxisArrow = UtilSp3dGeometry.cone(arrowHeadR, arrowHeadH,
           material: materialZ ?? FSp3dMaterial.blueNonWire.deepCopy());
@@ -122,11 +119,11 @@ class UtilSp3dCommonParts {
     gridBackFront.move(Sp3dV3D(size / 2, size / 2, 0));
     Sp3dObj gridFloorFront = gridBackFront
         .deepCopy()
-        .rotate(Sp3dV3D(1, 0, 0), 90 * _toRadian)
+        .rotate(Sp3dV3D(1, 0, 0), 90 * Sp3dConstantValues.toRadian)
         .reversed();
     Sp3dObj gridLeftFront = gridBackFront
         .deepCopy()
-        .rotate(Sp3dV3D(0, 1, 0), -90 * _toRadian)
+        .rotate(Sp3dV3D(0, 1, 0), -90 * Sp3dConstantValues.toRadian)
         .reversed();
     objList.add(gridBackFront);
     objList.add(gridFloorFront);