+++ lib/util_simple_3d.dart
@@ -1,3 +1,7 @@
+export 'src/math/sp3d_constant_values.dart';
+export 'src/math/util_bezier.dart';
+export 'src/math/util_sp3d_list.dart';
+export 'src/math/vrange.dart';
 export 'src/f_sp3d_material.dart';
 export 'src/util_sp3d_geometry.dart';
 export 'src/util_sp3d_common_parts.dart';